+++ src/Framework/Framework/Kernel.php
@@ -68,6 +68,8 @@ abstract class Kernel extends AbstractKernel
      * Get list of defined application bootloaders
      *
      * @return array<int, class-string>|array<class-string, array<non-empty-string, mixed>>
+     *
+     * @deprecated since v3.10 Use {@see defineBootloaders()} instead. Will be removed in v4.0
      */
     protected function defineAppBootloaders(): array
     {